implementation for get the most priority register and fix in priceDTO price data type

+++ src/main/java/com/comerce/comerce/repository/models/PriceDTO.java
@@ -47,7 +47,7 @@ public class PriceDTO implements Serializable {
 
     @NotEmpty(message = "cost may not be empty")
     @Column(name = "price")
-    private Long cost;
+    private Float cost;
 
     @Column(name = "curr")
     @NotEmpty(message = "currencyIso may not be empty")
